Fixing instagram login
When we have more than one header, for instance, two 'Set-Cookie'
headers,
they are concatenated using ",", not ";". This generates a parser error
since this program is expecting a string like "key=value;key=value;..."

It is not possible to know where the "," was added since it is a valid
character for values. The dirt fix below will solve the problem while
instagram does not change the values inside cookies